Abrex 400
Vandan Steel & Engg. co. is a ABREX 400 Plate importers, suppliers and stockists, conveying to the entire of the world.We are having biggest stock and prepared load of around 1000 tons of ABREX 400 Plate in various size, grades and cuts.Our fabricating system is upheld by escalated in-house item assessment and testing and authorize to meet all the most perceived global principles, including ISO 9001:2008

https://www.wearplates.co.in